Allow NixpkgsPlatformGroup, use distribution-nixpkgs for parsing
This change no longer exclusively produces `NixpkgsPlatformSinge`s. The
main hurdle for this was parsing: We need to be able to denote platform
groups in the yaml configuration somehow. nixpkgsPlatformFromString
which was added in <https://github.com/NixOS/distribution-nixpkgs/pull/19>
fills in this gap: Any string prefixed with either `platforms.` or
`lib.platforms.` will be interpreted as a platform group, all other
strings will (roughly) be treated as before.

Since the platform parsing capabilities of distribution-nixpkgs no
longer rely on a hard-coded list of known platforms, we can also reuse
it to parse the `--system` and `--platform` flags. In doing so, we
replace `parsePlatform` with `parseSinglePlatform` which is now also
reused in `hackage2nix`:

* Parsing of Nix-style platform tuples is handled by
  `nixpkgsPlatformFromString` (`NixpkgsPlatformGroup` results are
  rejected, of course).

* Parsing of LLVM-style triples is now handled by Cabal's
  `platformFromTriple`. Its implementation is practically equivalent to
  the `parsePlatformParts` case matching lists with more than two
  elements, so this should cause no problems.

* We loose the `ghcConvertArch` and `ghcConvertOS` functions. Their
  functionality is *partially* implemented in
  `nixpkgsPlatformFromString`, but not at all in
  `platformFromTriple`. This means that platform parsing will behave
  differently in edge cases (specifically when platform tuples / triples
  are used that are unsupported by Cabal, but known and fixed up in
  GHC's make build system). Consequently this is technically a breaking
  change, but in practice probably doesn't affect a lot of people.